Mobile Buttons, Save Level Progress
Added mobile buttons. These are a set of buttons that toggle a boolean when pressed, which can used in the Update function. I had a couple other ideas on how to do this, but this tied in the easiest with the existing code.
Save level progress.  JSON file is created on the operating system's Persistent Data Path, with an int of their furthest progress. This is read when opening up the game, and written to whenever the highestlevel is being changed.