AdMob update, UI asset cleanup
Updated AdMob systems by deleting and installing the new package

Got Interstitials mostly working. Basically, when going to the main menu, call up an interstitial, wait for the player to close/for it to fail to load, then actually go to the main menu.
Thing is, this doesn't really work with the current AdMob embargo on my account, so we'll comment it out for now.

Deleted more of the UI assets I downloaded that aren't Jesse's

Got the animated bullets started

Level 8 remade. Now "Lockpicking", a level that you must go down the middle of the level, even though it looks like you can go around