Title Screen Splash
Created animations for moving between the two main title screens. This is a Animation Controller attached to an empty gameobject that holds all of the gameobjects included (for the splash, the two texts, and for the buttons, the four buttons), and given an animation that moves between its start position and its end position.

This system should be extended and generified for the different pop up screens for the menu buttons.